move tchap scripts to tchap subfolder
