+++ .github/workflows/cypress.yaml
@@ -1,6 +1,6 @@
 # Run the tests
 # Todo : reuse a build from another action, instead of rebuilding for nothing.
-# tchap modification : use "./scripts/install-yarn-linked-repositories.sh" instead of "./scripts/layered.sh"
+# tchap modification : use "./scripts/tchap/install-yarn-linked-repositories.sh" instead of "./scripts/layered.sh"
 name: Cypress End to End Tests
 on:
     pull_request: {}
@@ -28,7 +28,7 @@ jobs:
                   node-version: 16
 
             - name: Install Dependencies
-              run: "./scripts/install-yarn-linked-repositories.sh"
+              run: "./scripts/tchap/install-yarn-linked-repositories.sh"
 
             - name: Run Cypress tests
               uses: cypress-io/github-action@v4.1.1
