+++ .github/workflows/static_analysis.yaml
@@ -1,5 +1,5 @@
 #Run a static analysis
-#tchap modification : use "./scripts/install-yarn-linked-repositories.sh" instead of "./scripts/layered.sh"
+#tchap modification : use "./scripts/tchap/install-yarn-linked-repositories.sh" instead of "./scripts/layered.sh"
 name: Static Analysis
 on:
     pull_request: {}
@@ -27,7 +27,7 @@ jobs:
     #          node-version: 16
     #
     #      - name: Install Dependencies
-    #        run: "./scripts/install-yarn-linked-repositories.sh"
+    #        run: "./scripts/tchap/install-yarn-linked-repositories.sh"
     #
     #      - name: Typecheck
     #        run: "yarn run lint:types"
@@ -68,7 +68,7 @@ jobs:
 
             # Needs branch matching as it inherits .stylelintrc.js from matrix-react-sdk
             - name: Install Dependencies
-              run: "./scripts/install-yarn-linked-repositories.sh"
+              run: "./scripts/tchap/install-yarn-linked-repositories.sh"
 
             - name: Run Linter
               run: "yarn run lint:style"
@@ -85,7 +85,7 @@ jobs:
 #         node-version: 16
 
 #     - name: Install Deps
-#       run: "scripts/install-yarn-linked-repositories.sh"
+#       run: "scripts/tchap/install-yarn-linked-repositories.sh"
 
 #     - name: Dead Code Analysis
 #       run: "yarn run analyse:unused-exports"
